Initial copy from gregli/decimal

+++ src/tests/Microsoft.PowerFx.Interpreter.Tests/PADIntegrationTests.cs
@@ -67,11 +67,11 @@ namespace Microsoft.PowerFx.Tests
             var result1 = engine.Eval("Value(Index(robintable, 1).Column1)"); // 101
             var result2 = engine.Eval("Text(Index(robintable, 2).Column2)"); // "str202"
 
-            Assert.Equal(101.0, result1.ToObject());
+            Assert.Equal(101m, result1.ToObject());
             Assert.Equal("str202", result2.ToObject());
 
             var result3 = engine.Eval("Sum(robintable, Value(ThisRecord.Column1))");
-            Assert.Equal(101.0 + 201 + 301, result3.ToObject());
+            Assert.Equal(101m + 201 + 301, result3.ToObject());
         }
 
         // Create table with strong typing
@@ -79,7 +79,7 @@ namespace Microsoft.PowerFx.Tests
         {
             var table = new DataTable();
 
-            table.Columns.Add("Scores", typeof(int));
+            table.Columns.Add("Scores", typeof(decimal));
             table.Columns.Add("Names", typeof(string));
 
             table.Rows.Add(10, "name1");
@@ -108,11 +108,11 @@ namespace Microsoft.PowerFx.Tests
             var result1 = engine.Eval("Index(robintable, 2).Scores"); // 20
             var result2 = engine.Eval("Index(robintable, 3).Names"); // "name3"
 
-            Assert.Equal(20.0, result1.ToObject());
+            Assert.Equal(20m, result1.ToObject());
             Assert.Equal("name3", result2.ToObject());
 
             var result3 = engine.Eval("Sum(robintable, ThisRecord.Scores)");
-            Assert.Equal(60.0, result3.ToObject());
+            Assert.Equal(60m, result3.ToObject());
 
             // Access field not on the table 
             var result4 = engine.Eval(@"
@@ -132,7 +132,7 @@ First(
 
             Assert.Equal(3, table.Rows.Count);
 
-            var result5 = engine.Eval("Remove(robintable, {Names:\"name2\"});robintable", options: opt);            
+            var result5 = engine.Eval("Remove(robintable, {Names:\"name2\"});robintable", options: opt);
             Assert.Equal("Table({Names:\"name1\",Scores:10},{Names:\"name3\",Scores:30})", ((DataTableValue)result5).Dump());
 
             // Is table object affected?
@@ -186,7 +186,7 @@ First(
             var result1 = engine.Eval("Value(Index(robinList, 1).Value)");
             var result2 = engine.Eval("Text(Index(robinList, 2).Value)");
 
-            Assert.Equal(1.0, result1.ToObject());
+            Assert.Equal(1m, result1.ToObject());
             Assert.Equal("string", result2.ToObject());
         }
     }
